Apply remarks
