+++ dpnp/tests/test_arraycreation.py
@@ -56,33 +56,6 @@ class TestArray:
         assert_raises(TypeError, dpnp.array, x, ndmin=3.0)
 
 
-class TestAttributes:
-    def setup_method(self):
-        self.one = dpnp.arange(10)
-        self.two = dpnp.arange(20).reshape(4, 5)
-        self.three = dpnp.arange(60).reshape(2, 5, 6)
-
-    def test_attributes(self):
-        assert_equal(self.one.shape, (10,))
-        assert_equal(self.two.shape, (4, 5))
-        assert_equal(self.three.shape, (2, 5, 6))
-        self.three.shape = (10, 3, 2)
-        assert_equal(self.three.shape, (10, 3, 2))
-        self.three.shape = (2, 5, 6)
-        assert_equal(self.one.strides, (self.one.itemsize / self.one.itemsize,))
-        num = self.two.itemsize / self.two.itemsize
-        assert_equal(self.two.strides, (5 * num, num))
-        num = self.three.itemsize / self.three.itemsize
-        assert_equal(self.three.strides, (30 * num, 6 * num, num))
-        assert_equal(self.one.ndim, 1)
-        assert_equal(self.two.ndim, 2)
-        assert_equal(self.three.ndim, 3)
-        num = self.two.itemsize
-        assert_equal(self.two.size, 20)
-        assert_equal(self.two.nbytes, 20 * num)
-        assert_equal(self.two.itemsize, self.two.dtype.itemsize)
-
-
 class TestTrace:
     @pytest.mark.parametrize("a_sh", [(3, 4), (2, 2, 2)])
     @pytest.mark.parametrize(
@@ -241,20 +214,6 @@ def test_arange(start, stop, step, dtype):
         assert_array_equal(exp_array, res_array)
 
 
-@pytest.mark.parametrize(
-    "arr",
-    [
-        numpy.array([1]),
-        dpnp.array([1]),
-        [1],
-    ],
-    ids=["numpy", "dpnp", "list"],
-)
-def test_create_from_usm_ndarray_error(arr):
-    with pytest.raises(TypeError):
-        dpnp.dpnp_array.dpnp_array._create_from_usm_ndarray(arr)
-
-
 @pytest.mark.parametrize("func", ["diag", "diagflat"])
 @pytest.mark.parametrize("k", [-6, -5, -4, -3, -2, -1, 0, 1, 2, 3, 4, 5, 6])
 @pytest.mark.parametrize(
@@ -384,8 +343,7 @@ def test_identity(n, dtype):
 
 def test_identity_error():
     # negative dimensions
-    with pytest.raises(ValueError):
-        _ = dpnp.identity(-5)
+    assert_raises(ValueError, dpnp.identity, -5)
 
 
 @pytest.mark.parametrize("dtype", get_all_dtypes(no_float16=False))
